requirements.txt
# Core bot framework
aiogram==3.4.1

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Environment variables
python-dotenv==1.0.0

# Logging (built-in, but for completeness)
# logging - built into Python

# Development dependencies (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1